+++ src/io_blowup.F90
@@ -268,7 +268,7 @@ MODULE io_BLOWUP
 		character(len=*), intent(in)           :: name
 		integer, intent(in)                    :: dims(1)
 		character(len=*), intent(in), optional :: units, longname
-		real(kind=8),target,     intent(inout)        :: data(:)
+		real(kind=WP),target,     intent(inout)        :: data(:)
 		integer                                :: c
 		type(nc_vars), allocatable, dimension(:) :: temp
 		
@@ -304,7 +304,7 @@ MODULE io_BLOWUP
 		character(len=*), intent(in)           :: name
 		integer, intent(in)                    :: dims(2)
 		character(len=*), intent(in), optional :: units, longname
-		real(kind=8),target,     intent(inout) :: data(:,:)
+		real(kind=WP),target,     intent(inout) :: data(:,:)
 		integer                                :: c
 		type(nc_vars), allocatable, dimension(:) :: temp
 		
@@ -338,7 +338,7 @@ MODULE io_BLOWUP
 		implicit none
 		type(nc_file),  intent(inout) :: id
 		integer,  intent(in)          :: istep
-		real(kind=8), allocatable     :: aux1(:), aux2(:,:) 
+		real(kind=WP), allocatable     :: aux1(:), aux2(:,:) 
 		integer                       :: i, size1, size2, shape
 		integer                       :: c
 		! Serial output implemented so far
@@ -396,7 +396,7 @@ MODULE io_BLOWUP
 		type(nc_file),  intent(inout) :: id
 		character(500)                :: longname
 		integer                       :: c, j, k
-		real(kind=8)                  :: rtime !timestamp of the record
+		real(kind=WP)                 :: rtime !timestamp of the record
 		! Serial output implemented so far
 		if (mype/=0) return
 		c=1
